telemetry: extract transition metrics module
