+++ apps/backend/app/domains/navigation/api/preview_router.py
@@ -13,8 +13,8 @@ from sqlalchemy.future import select
 from app.core.db.session import get_db
 
 from app.core.preview import PreviewContext, PreviewMode  # isort: skip
-from app.core.metrics import record_no_route, record_route_length
 from app.core.rng import next_seed
+from app.core.transition_metrics import record_no_route, record_route_length
 from app.domains.navigation.application.navigation_service import NavigationService
 from app.domains.navigation.application.transition_router import (
     NoRouteReason,
