+++ apps/backend/app/domains/telemetry/api/admin_metrics_router.py
@@ -5,7 +5,8 @@ from typing import Annotated
 from fastapi import APIRouter, Depends, Query
 from pydantic import BaseModel
 
-from app.core.metrics import _percentile, metrics_storage, transition_stats
+from app.core.metrics import _percentile, metrics_storage
+from app.core.transition_metrics import transition_stats
 from app.domains.telemetry.application.event_metrics_facade import event_metrics
 from app.security import ADMIN_AUTH_RESPONSES, require_admin_role
